+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 + Implement `everyNth()` to get every `n`<sup>th</sup> element from the stream
 + Implement `uniquelyOccurring()` to emit stream elements that occur a single time
 + Implement `takeUntil()` to take from a stream until a predicate is met, including the first element that matches the predicate
-+ Implement `foldIndexed()` to perform a traditional fold along with the index of each element
-+ Implement `reduceIndexed()` to perform a reduce along with the index of each element
++ Implement `foldIndexed()` to perform a fold along with the index of each element
++ Implement `scanIndexed()` to perform a scan along with the index of each element
 
 ### 0.7.0
 + Use greedy integrators where possible (Fixes #57)